ai-server/requirements.txt
# AI Server Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
deepface==0.0.86
opencv-python-headless==4.8.1.78
numpy==1.24.3
tensorflow==2.13.0
python-multipart==0.0.6
Pillow==10.1.0
scipy==1.11.4
mtcnn==0.1.1
retina-face==0.0.14
requests==2.31.0
python-jose==3.3.0
